web-backend/requirements.txt
boto3
botocore
Django
django-filter
django-db-views
djangorestframework
djangorestframework-simplejwt
PyJWT
drf-yasg
bcrypt
celery
django-apscheduler
django-storages
boto3
pydantic
python-environ
django-celery-results
pylint_django
gunicorn
drf-spectacular
channels[daphne]
selenium
channels_redis
pycryptodome
langchain
langchain_openai
langchain_community
faker
pillow
psycopg2-binary
pgvector
tiktoken